tech/backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arrow==1.2.3
binaryornot==0.4.4
certifi==2024.6.2
chardet==5.1.0
click==8.1.3
click-default-group==1.2.4
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jinja2-time==0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
paho-mqtt==2.1.0
pip==24.0
pluggy==1.5.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.1
PyYAML==6.0
rich==13.7.1
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-utils==3.36
starlette==0.37.2
tabulate==0.9.0
text-unidecode==1.3
typer==0.12.3
typing_extensions==4.12.1
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wheel==0.43.0
